changes testing based

+++ Fuzzy_dht_NIOT/Fuzzy_dht_NIOT.ino
@@ -86,7 +86,7 @@ void loop()
   // Running the Fuzzification
   fuzzy->fuzzify();
   // Running the Defuzzification
-  float output = fuzzy->defuzzify(1);
+  int output = fuzzy->defuzzify(1);
   // Printing something
   Serial.println("Result: ");
   Serial.print("\t\t\tTempSet: ");
